backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.28.0
redis>=5.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
passlib[bcrypt]>=1.7.4
pyjwt>=2.8.0
python-multipart>=0.0.6
alembic>=1.12.0
celery[redis]>=5.3.4
openai>=1.14.0
tenacity>=8.2.3
groq>=0.4.1

